requirements.txt
pydantic==2.9.2
# valo-api[async]==2.0.8
ratelimit==2.2.1
psycopg2-binary==2.9.10
tqdm
httpx==0.27.2
python-dotenv==0.19.1
numpy
scikit-learn
pandas
seaborn
nltk
matplotlib
statsmodels